CSCEXAM-000 changes for DI wiring - some random timing problem when resolving Ebean server during startup - move EnrolmentRepository injection away from BaseController (not really needed there) - use an injection provider for EbeanConfig and delay resolving EbeanServer - hopefully this fixes the problems starting in prod mode (in some environments)
